changed uint to uint256

+++ contracts/Factories/CErc20ImmutableFactory.sol
@@ -8,7 +8,7 @@ contract CErc20ImmutableFactory {
     address underlying_,
     ComptrollerInterface comptroller_,
     InterestRateModel interestRateModel_,
-    uint initialExchangeRateMantissa_,
+    uint256 initialExchangeRateMantissa_,
     string memory name_,
     string memory symbol_,
     uint8 decimals_,
